move `GetFieldVectorShape` to internal
